extract controller creation to game.ts
