+++ src/game/cyberball-game-model.ts
